Feature/docs for0.1.0 (#13)
* Add swift docs

* Fix: remove .zero from MoveMouse in Example project
