+++ Sources/SwiftRobot/Capabilities/SwiftRobotScreen.swift
@@ -7,19 +7,23 @@
 
 import CoreGraphics
 
+/// A component capable of using RobotScreen to get information of the screen
 public struct SwiftRobotScreen: RobotScreenCapable {
+    /// Get the bounds of the display
     public func bounds(
         display: CGDirectDisplayID = CGMainDisplayID()
     ) async -> CGRect {
         await screen.bounds(display: display)
     }
     
+    /// Get the center point of the bounds of the display
     public func center(
         display: CGDirectDisplayID = CGMainDisplayID()
     ) async -> CGPoint {
         await screen.center(display: display)
     }
     
+    /// Create an image of the display for a certain CGRect
     public func image(
         display: CGDirectDisplayID = CGMainDisplayID(),
         rect: CGRect
