+++ Sources/SwiftRobot/Robot+SwiftRobot.swift
@@ -11,6 +11,7 @@ public enum SwiftRobot { }
 
 // MARK: - Static Properties
 
+///
 public extension SwiftRobot {
     static var screen: SwiftRobotScreen = SwiftRobotScreen()
     static var mouse: SwiftRobotMouse = SwiftRobotMouse()
@@ -21,13 +22,17 @@ public extension SwiftRobot {
 
 // MARK: - Static Functions
 
+/// SwiftRobot runnable functionatlity
 public extension SwiftRobot {
+    
+    /// to run one RoboTask asynchronously
     static func run(
         @RobotTaskBuilder _ builder: () async throws -> RobotTask
     ) async throws {
         try await builder().run()
     }
     
+    /// to run one or more RoboTask asynchronously
     static func run(
         @RobotTaskBuilder _ builder: () async throws -> [RobotTask]
     ) async throws {
@@ -36,6 +41,7 @@ public extension SwiftRobot {
         }
     }
     
+    /// to run one RoboTask asynchronously on the Main Thread
     @MainActor
     static func main(
         @RobotTaskBuilder _ builder: () async throws -> RobotTask
@@ -43,6 +49,7 @@ public extension SwiftRobot {
         try await run(builder)
     }
     
+    /// to run one or more RoboTask asynchronously on the Main Thread
     @MainActor
     static func main(
         @RobotTaskBuilder _ builder: () async throws -> [RobotTask]
